Ensure only a single finder window exists
Launching a new file or buffer finder via a mapping while one exists
results in improper window height calculations.

In order to avoid this, we ensure that only a single active finder
exists: Activating a finder while the same kind is open hides the
finder; activating a finder different from the one open hides the active
finder and opens the requested one.

A side effect of aggressively hiding the finder is that there may not
exist a buffer to unload/wipeout when closing a window. Executing with
`silent!' silences the unneeded error message.