Set synmaxcol to avoid match highlighting errors
The new match highlight feature uses <commandt></commandt> as conceal
markers per match. If the user has set a low synmaxcol (in order to
speed up syntax highlighting in files with long lines), the concealed
markers can quickly bring the character count of the line past the
synmaxcol value, causing all markers past the value to be shown.

This can be demonstrated by setting a low synmaxcol, then opening a new
match window in this project's root:

:set synmaxcol=20 | CommandT

" Search for `//`

> fixtures/bar/abc
  fixtures/</commandt>bar<commandt>/</commandt>xyz
  fixtures/</commandt>foo<commandt>/</commandt>beta